Add separate operator* functions for Point Vector
Just having one function that takes a Tuple and returns a Tuple is not enough.
Sometimes you want to assign that returning Tuple to a Vector, and
there is no conversion from Tuple to Vector.

So I created two new functions, bascially replicas of the Tuple version.

I wish I knew of there was a better way to express this.